- improvements for bird projects - additional checkbox in script inventory to populate script menue - filters in filter drop down are now alphabetically ordered - more information in log in case of errors while AI is running - email with last logs in help menue

+++ BatInspector/ViewModel/ZoomView.cs
@@ -480,7 +480,10 @@ namespace BatInspector
               dstFile = Path.Combine(dstDir, Path.GetFileName(srcFile));
             else
               dstFile = Path.Combine(dstDir, prefix + "_" + Path.GetFileName(srcFile));
-            File.Copy(srcFile, dstFile);
+            if (File.Exists(srcFile))
+              File.Copy(srcFile, dstFile);
+            else
+              DebugLog.log($"PNG file: {srcFile} doesn'exist!", enLogType.WARNING);
           }
           if (incXml)
           {
@@ -489,7 +492,10 @@ namespace BatInspector
               dstFile = Path.Combine(dstDir, Path.GetFileName(srcFile));
             else
               dstFile = Path.Combine(dstDir, prefix + "_" + Path.GetFileName(srcFile));
-            File.Copy(srcFile, dstFile);
+            if (File.Exists(srcFile))
+              File.Copy(srcFile, dstFile);
+            else
+              DebugLog.log($"XML file: {srcFile} doesn'exist!", enLogType.WARNING);
           }
         }
         else
